mark5access: add support for >8 Gbps/channel via a change of mark5_stream.samprate from int to int64_t (also tested that mark5access examples and difx worked correctly for 16 Gbps VDIF just by recompiling them)
git-svn-id: https://svn.atnf.csiro.au/difx@6929 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/mark5access/trunk/mark5access/format_vdifb.c
@@ -2099,7 +2099,7 @@ static int mark5_format_vdifb_init(struct mark5_stream *ms)
 			framensNum, framensDen);
 			ms->framegranularity = 1;
 		}
-		ms->samprate = ms->framesamples*(1000000000.0/ms->framens);
+		ms->samprate = ((int64_t)ms->framesamples)*(1000000000.0/ms->framens);
         }
         else
         {
